Workaround part being considered 'changed' by snapcraft 2.x
There's an known issue of Snapcraft 2.x that parts with `nil` plugin
will implicitly have `source: .`, which causes the part to be considered
'changed' when files under `.` is modified and thus cleaning all the
depending parts.

This patch implements a fake source folder to workaround this behavior.

Thanks for Olivier Tilloy for the workaround.

Refer-to: Chromium fails to build on !amd64: snapcraft complains that
the 'pull' step of 'gtk3-locales' is out of date - snapcraft -
snapcraft.io
<https://forum.snapcraft.io/t/chromium-fails-to-build-on-amd64-snapcraft-complains-that-the-pull-step-of-gtk3-locales-is-out-of-date/7655/13>
Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>